requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
Click==7.0
Flask==1.1.1
html2text==2019.8.11
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
psutil==5.6.7
py4j==0.10.7
pyspark==2.4.4
soupsieve==1.9.5
Werkzeug==0.16.0
wincertstore==0.2
requests
Cython
numpy
azure-cosmos==3.1.2